refactor: add dbOptions to separate config file
