+++ product-service/src/functions/createProduct/handler.ts
@@ -6,21 +6,7 @@ import { formatJSONResponse } from '@libs/apiGateway';
 import { middyfy } from '@libs/lambda';
 
 import schema from './schema';
-// import { getAllProducts } from '../../repository/products-repository';
-
-const {PG_HOST, PG_PORT, PG_DBNAME, PG_USERNAME, PG_PASSWORD} = process.env;
-
-const dbOptions = {
-  host: PG_HOST,
-  port: PG_PORT,
-  database: PG_DBNAME,
-  user: PG_USERNAME,
-  password: PG_PASSWORD,
-  ssl:{
-    rejectUnauthorized: false
-  },
-  connectionTimeoutMillis: 5000
-};
+import { dbOptions } from '../../config/database-config';
 
 const createProduct: ValidatedEventAPIGatewayProxyEvent<typeof schema> = async (event) => {
     console.log("createProduct lambda launched");
