+++ product-service/src/functions/getProductsById/handler.ts
@@ -6,27 +6,12 @@ import { formatJSONResponse } from '@libs/apiGateway';
 import { middyfy } from '@libs/lambda';
 
 import schema from './schema';
-// import { findProductById } from '../../repository/products-repository';
-
-const {PG_HOST, PG_PORT, PG_DBNAME, PG_USERNAME, PG_PASSWORD} = process.env;
-
-const dbOptions = {
-  host: PG_HOST,
-  port: PG_PORT,
-  database: PG_DBNAME,
-  user: PG_USERNAME,
-  password: PG_PASSWORD,
-  ssl:{
-    rejectUnauthorized: false
-  },
-  connectionTimeoutMillis: 5000
-};
+import { dbOptions } from '../../config/database-config';
 
 const getProductsById: ValidatedEventAPIGatewayProxyEvent<typeof schema> = async (event) => {
   console.log("getProductsById lambda launched with event: ", event);
   console.log("Product Id: ", event.pathParameters.productId);
 
-  // /^[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$/i - all versions
   if( !/^[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$/i.test(event.pathParameters.productId) ){
     console.log("Invalid v4 UUID provided");
