+++ product-service/src/functions/getProductsList/handler.ts
@@ -1,26 +1,12 @@
 import 'source-map-support/register';
-import {Client} from 'pg';
+import { Client } from 'pg';
 
 import type { ValidatedEventAPIGatewayProxyEvent } from '@libs/apiGateway';
 import { formatJSONResponse } from '@libs/apiGateway';
 import { middyfy } from '@libs/lambda';
 
 import schema from './schema';
-// import { getAllProducts } from '../../repository/products-repository';
-
-const {PG_HOST, PG_PORT, PG_DBNAME, PG_USERNAME, PG_PASSWORD} = process.env;
-
-const dbOptions = {
-  host: PG_HOST,
-  port: PG_PORT,
-  database: PG_DBNAME,
-  user: PG_USERNAME,
-  password: PG_PASSWORD,
-  ssl:{
-    rejectUnauthorized: false
-  },
-  connectionTimeoutMillis: 5000
-};
+import { dbOptions } from '../../config/database-config';
 
 const getProductsList: ValidatedEventAPIGatewayProxyEvent<typeof schema> = async () => {
     console.log("getProductsList lambda launched");
@@ -40,7 +26,7 @@ const getProductsList: ValidatedEventAPIGatewayProxyEvent<typeof schema> = async
         message: "failed to fetch data"
       });
     }finally{
-      client.end();
+      await client.end();
     }
 }
